Fix PR review issues: overflow, hardcoded skills, caching, error handling
1. CRITICAL: Change profDiversity from uint8_t to uint32_t in ProductionSystem
   to prevent overflow when ProfessionID >= 8.

2. CRITICAL: Generalize skill rust in AgentDecisionSystem to iterate over all
   professions from the schema instead of hardcoding RES_FOOD/WATER/WOOD.

3. SERIOUS: Cache idleProfessionId and haulerProfessionId in WorldSchema,
   populated in BuildMaps(), making FindIdleProfession/FindHaulerProfession O(1).

4. SERIOUS: ProfessionForResource now returns INVALID_ID instead of the idle
   profession when no mapping exists. Callers in BirthSystem and
   AgentDecisionSystem updated to handle INVALID_ID gracefully.

5. SERIOUS: Add assertions and stderr logging in WorldGenerator when
   FindProfession returns INVALID_ID for required professions.

6. SERIOUS: Remove deprecated ProfessionType enum from Components.h.

+++ src/ECS/Systems/ProductionSystem.cpp
@@ -58,7 +58,7 @@ void ProductionSystem::Update(entt::registry& registry, float realDt, const Worl
     // [settlement][resourceIndex 0=Food,1=Water,2=Wood]
     std::unordered_map<entt::entity, std::array<SkillAccum, 3>> skillData;
     // Profession diversity: bitmask per settlement (bit N = professionID N that produces a resource)
-    std::unordered_map<entt::entity, uint8_t> profDiversity;
+    std::unordered_map<entt::entity, uint32_t> profDiversity;
 
     auto resIdx = [](int rt) -> int {
         switch (rt) {
